feat(routes): swap homepage to classic form, move conversation hero to /ai, /form->/ redirect
